Refactor out typeDefs to separate files
