+++ api/graphql/typeDefs/misc.ts
@@ -1,0 +1,7 @@
+export default `
+    type User {
+        email: String!
+    }
+    
+    scalar Date
+`;
